implemented initial panels
