+++ client/src/main/java/client/Main/ClientMain.java
@@ -1,14 +1,29 @@
-package client;
+package client.Main;
 
-import javax.swing.*;
-import java.awt.*;
+// 외부 import
+import java.awt.BorderLayout;
+import java.awt.Color;
+import java.awt.Component;
+import java.awt.Dimension;
+import java.awt.Font;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.awt.GridLayout;
+
+
+import javax.swing.Box;
+import javax.swing.BoxLayout;
+import javax.swing.JButton;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+import javax.swing.JTextField;
+
+// 내부 import
+import client.Frame;
+import client.Main.StatisticsPanel;
 
 public class ClientMain extends JPanel {
-  private JPanel currentPanel;
-
-  public ClientMain() {
+  public ClientMain(Frame parentFrame) {
 
     // 바텀 내비게이션 바
     JPanel bottomNavBar = new JPanel();
@@ -28,38 +43,13 @@ public class ClientMain extends JPanel {
     JButton profileButton = new JButton("프로필");
     profileButton.setPreferredSize(new Dimension(50, 50));
 
-    // 버튼 클릭 이벤트 추가
-    // 버튼 클릭 이벤트 추가
-    ActionListener buttonListener = new ActionListener() {
+    // signUpButton 클릭 이벤트
+    statisticsButton.addActionListener(new ActionListener() {
       @Override
       public void actionPerformed(ActionEvent e) {
-        // 현재 패널을 제거하고 새로운 패널로 교체
-        remove(currentPanel);
-
-        if (e.getSource() == homeButton) {
-          currentPanel = new HomePanel();
-        } else if (e.getSource() == goalButton) {
-          currentPanel = new GoalPanel();
-        } else if (e.getSource() == profileButton) {
-          currentPanel = new ProfilePanel();
-        } else if (e.getSource() == statisticsButton) {
-          currentPanel = new StatisticsPanel();
-        } else if (e.getSource() == listButton) {
-          currentPanel = new ListPanel();
-        }
-
-        // 새로운 패널 추가 및 화면 갱신
-        add(currentPanel, BorderLayout.CENTER);
-        revalidate();
-        repaint();
+        parentFrame.switchToPanel(new StatisticsPanel(parentFrame));
       }
-    };
-
-    homeButton.addActionListener(buttonListener);
-    profileButton.addActionListener(buttonListener);
-    statisticsButton.addActionListener(buttonListener);
-    goalButton.addActionListener(buttonListener);
-    listButton.addActionListener(buttonListener);
+    });
 
     // 버튼을 바텀 내비게이션 바에 추가
     bottomNavBar.add(statisticsButton);
