+++ client/src/main/java/client/Main/GoalPanel.class
