+++ client/src/main/java/client/Main/GoalPanel.java
@@ -1,7 +1,11 @@
-package client;
+package client.Main;
 
-import javax.swing.*;
-import java.awt.*;
+import java.awt.BorderLayout;
+import java.awt.Font;
+
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+import javax.swing.SwingConstants;
 
 public class GoalPanel extends JPanel {
   public GoalPanel() {
