+++ client/src/main/java/client/Main/HomePanel.java
@@ -1,8 +1,17 @@
-package client;
+package client.Main;
 
-import javax.swing.*;
+import java.awt.BorderLayout;
+import java.awt.Color;
+import java.awt.Component;
+import java.awt.Dimension;
+import java.awt.Font;
+
+import javax.swing.Box;
+import javax.swing.BoxLayout;
+import javax.swing.JCheckBox;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
 import javax.swing.border.LineBorder;
-import java.awt.*;
 
 public class HomePanel extends JPanel {
     public HomePanel() {
